create dockertemplate class for erb processing
This change adds DockerTemplate class so we can use
it for generic erb template class processing.   We will
use this class on forj-docker cli as well later so
that builds can include erb templates for docker files.

When implemented we should add a dockerfile defaults
section to the configure command so that we can generate
generic config for all dockerfiles that use erb templates.

* remove rubocop checks in tmp folder

Change-Id: I0d5eff72d5eb842e74e2441c282d364027a9323b