implement forj-docker template command
This command can be used durring the docker_prepare.sh
step to convert a Docker template to a Dockerfile, but
first we must implement forj-docker set and forj-docker get
so we can configure default key value pairs to use
in a docker template beyond what comes in the default.

We will aslo use this to make a translation of
meta_data that should be stored in meta.js at
start time.

Change-Id: Id21b1f9aad9058e92188eec6db843135d1332a5d